% Generated by maven-r-jsr223-plugin: do not edit by hand

\name{CitationGraphAnalyser}

\alias{CitationGraphAnalyser}

\title{}

\usage{\preformatted{
library(jlitrev)
J = JavaApi$new();
instance <- J$CitationGraphAnalyser$new(...);
}}

\description{
	
}

\details{
	
}

\examples{
## -----------------------------------
## Method `CitationGraphAnalyser$setReferenceFormat`
## -----------------------------------

## -----------------------------------
## Method `CitationGraphAnalyser$executeQuery`
## -----------------------------------

## -----------------------------------
## Method `CitationGraphAnalyser$shutdown`
## -----------------------------------

## -----------------------------------
## Method `CitationGraphAnalyser$getQueryNames`
## -----------------------------------

}

\keyword{java api}

\section{Methods}{
	\subsection{Constructors}{
		\itemize{
			\item \href{#method-new}{\code{J$CitationGraphAnalyser$new(graph)}}
		}
	}
	\subsection{Static methods}{
		\itemize{
		}
	}
	\subsection{Instance methods}{
		\itemize{
			\item \href{#method-setReferenceFormat}{\code{instance$setReferenceFormat(referenceFormat)}}
			\item \href{#method-executeQuery}{\code{instance$executeQuery(qryName)}}
			\item \href{#method-shutdown}{\code{instance$shutdown()}}
			\item \href{#method-getQueryNames}{\code{instance$getQueryNames()}}
			\item \code{instance$clone()}
			\item \code{instance$print()}
		}
	}

	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-new"></a>}}
	
	\subsection{Method \code{new()}}{
		
	}

	\subsection{Usage}{
		\if{html}{\out{<div class="r">}}
		\preformatted{
  library(jlitrev)
  J <- JavaApi$new()
  J$CitationGraphAnalyser$new(graph)
		}
		\if{html}{\out{</div>}}
	}
	
	\subsection{Arguments}{
		\if{html}{\out{<div class="arguments">}}
		\describe{
			\itemize{
				\item{graph}{ - (java expects a uk.co.terminological.literaturereview.CitationGraph)}
			}
		}
		\if{html}{\out{</div>}}
	}

	\subsection{Returns}{
		R6 CitationGraphAnalyser object: 
	}

	\subsection{Examples}{
		\if{html}{\out{<div class="r example copy">}}
		\preformatted{
		}
		\if{html}{\out{</div>}}
	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-setReferenceFormat"></a>}}
	
	\subsection{Method \code{setReferenceFormat()}}{
		
	}

	\subsection{Usage}{
		\if{html}{\out{<div class="r">}}
		\preformatted{
  library(jlitrev)
  J <- JavaApi$new()
  instance <- J$CitationGraphAnalyser$new(...);
  instance$setReferenceFormat(referenceFormat)
		}
		\if{html}{\out{</div>}}
	}
	
	\subsection{Arguments}{
		\if{html}{\out{<div class="arguments">}}
		\describe{
			\itemize{
				\item{referenceFormat}{ - (java expects a java.lang.String)}
			}
		}
		\if{html}{\out{</div>}}
	}

	\subsection{Returns}{
		void: 
	}

	\subsection{Examples}{
		\if{html}{\out{<div class="r example copy">}}
		\preformatted{
		}
		\if{html}{\out{</div>}}
	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-executeQuery"></a>}}
	
	\subsection{Method \code{executeQuery()}}{
		
	}

	\subsection{Usage}{
		\if{html}{\out{<div class="r">}}
		\preformatted{
  library(jlitrev)
  J <- JavaApi$new()
  instance <- J$CitationGraphAnalyser$new(...);
  instance$executeQuery(qryName)
		}
		\if{html}{\out{</div>}}
	}
	
	\subsection{Arguments}{
		\if{html}{\out{<div class="arguments">}}
		\describe{
			\itemize{
				\item{qryName}{ - (java expects a java.lang.String)}
			}
		}
		\if{html}{\out{</div>}}
	}

	\subsection{Returns}{
		uk.co.terminological.jsr223.ROutput$Dataframe: 
	}

	\subsection{Examples}{
		\if{html}{\out{<div class="r example copy">}}
		\preformatted{
		}
		\if{html}{\out{</div>}}
	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-shutdown"></a>}}
	
	\subsection{Method \code{shutdown()}}{
		
	}

	\subsection{Usage}{
		\if{html}{\out{<div class="r">}}
		\preformatted{
  library(jlitrev)
  J <- JavaApi$new()
  instance <- J$CitationGraphAnalyser$new(...);
  instance$shutdown()
		}
		\if{html}{\out{</div>}}
	}
	
	\subsection{Arguments}{
		\if{html}{\out{<div class="arguments">}}
		\describe{
			\itemize{
			}
		}
		\if{html}{\out{</div>}}
	}

	\subsection{Returns}{
		void: 
	}

	\subsection{Examples}{
		\if{html}{\out{<div class="r example copy">}}
		\preformatted{
		}
		\if{html}{\out{</div>}}
	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-getQueryNames"></a>}}
	
	\subsection{Method \code{getQueryNames()}}{
		
	}

	\subsection{Usage}{
		\if{html}{\out{<div class="r">}}
		\preformatted{
  library(jlitrev)
  J <- JavaApi$new()
  instance <- J$CitationGraphAnalyser$new(...);
  instance$getQueryNames()
		}
		\if{html}{\out{</div>}}
	}
	
	\subsection{Arguments}{
		\if{html}{\out{<div class="arguments">}}
		\describe{
			\itemize{
			}
		}
		\if{html}{\out{</div>}}
	}

	\subsection{Returns}{
		java.util.List: 
	}

	\subsection{Examples}{
		\if{html}{\out{<div class="r example copy">}}
		\preformatted{
		}
		\if{html}{\out{</div>}}
	}
}