% Generated by r-jsr223-maven-plugin: do not edit by hand
\docType{package}
\name{jlitrev-package}
\alias{jlitrev}
\alias{jlitrev-package}
\title{jlitrev: R Java literature graph}
\description{
\if{html}{\figure{logo.png}{options: align='right'}}
A package to facilitate search and analysis of the literature

Version: 0.01

Classes:
\itemize{ 
	\item \code{JavaApi}
	\item \code{CitationGraph}
}
}

\author{
\itemize{ 
   \item Rob Challen rc538@exeter.ac.uk 
   \item r-jsr223-maven-plugin terminological  
}
}

\keyword{java api}