% Generated by maven-r-jsr223-plugin: do not edit by hand

\name{JavaApi}

\alias{JavaApi}

\title{}

\usage{\preformatted{
library(jlitrev)
J = JavaApi$new();
}}

\description{
	A package to facilitate search and analysis of the literature
}

\keyword{java api}

\section{Methods}{
	\subsection{Public methods}{
		\itemize{
			\item \code{JavaApi$new()}
			\item \href{#method-CitationGraph-new}{\code{J$CitationGraph$new(secrets, graphDb, graphConf, working, output)}}
			\item \href{#method-CitationGraph-fromPropertyFile}{\code{J$CitationGraph$fromPropertyFile(propFilename)}}
		}
	}

	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-new"></a>}}
	
	\subsection{Method \code{CitationGraph$new()}}{
		Loads a set of configuration parameters from a file and start up a neo4j graph.
	}

	\subsection{Usage}{
		\if{html}{\out{<div class="r">}}
		\preformatted{
  library(jlitrev)
  J <- JavaApi$new()
  J$CitationGraph$new(secrets, graphDb, graphConf, working, output)
		}
		\if{html}{\out{</div>}}
	}
	
	\subsection{Arguments}{
		\if{html}{\out{<div class="arguments">}}
		\describe{
			\itemize{
				\item{secrets}{ - (java expects a java.lang.String)}
				\item{graphDb}{ - (java expects a java.lang.String)}
				\item{graphConf}{ - (java expects a java.lang.String)}
				\item{working}{ - (java expects a java.lang.String)}
				\item{output}{ - (java expects a java.lang.String)}
			}
		}
		\if{html}{\out{</div>}}
	}

	\subsection{Returns}{
		R6 CitationGraph object: 
	}

	\subsection{Examples}{
		\if{html}{\out{<div class="r example copy">}}
		\preformatted{
		}
		\if{html}{\out{</div>}}
	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-fromPropertyFile"></a>}}
	
	\subsection{Method \code{CitationGraph$fromPropertyFile()}}{
		
	}

	\subsection{Usage}{
		\if{html}{\out{<div class="r">}}
		\preformatted{
  library(jlitrev)
  J <- JavaApi$new()
  J$CitationGraph$fromPropertyFile(propFilename)
		}
		\if{html}{\out{</div>}}
	}
	
	\subsection{Arguments}{
		\if{html}{\out{<div class="arguments">}}
		\describe{
			\itemize{
				\item{propFilename}{ - (java expects a java.lang.String)}
			}
		}
		\if{html}{\out{</div>}}
	}

	\subsection{Returns}{
		R6 CitationGraph object: 
	}

	\subsection{Examples}{
		\if{html}{\out{<div class="r example copy">}}
		\preformatted{
		}
		\if{html}{\out{</div>}}
	}
}